# openapi.yaml
openapi: 3.0.3
info:
  title: FreeSWITCH Call Control API
  description: |
    RESTful API for controlling FreeSWITCH calls and managing mod_callcenter
    resources (queues, agents, tiers, members) via Event Socket Library (ESL).

    ## Authentication

    When `FSAPI_AUTH_TOKENS` is configured, all requests (except from localhost)
    must include a valid Bearer token in the `Authorization` header.

    ## Multi-Tenant Authorization

    The `X-Allowed-Contexts` header enables per-tenant access control:
    - **Omitted or empty** — unrestricted access (backward compatible)
    - **`*`** — explicit unrestricted / super-admin access
    - **`domain1.com,domain2.com`** — restrict to listed domains

    For call endpoints the domain is matched against the call's `accountcode`.
    For callcenter endpoints it is matched against the domain portion of
    queue/tier names (`name@domain`) or the `domain_name=` value inside agent
    contact strings.
  version: 0.4.2
  contact:
    name: Emaktel
    url: https://github.com/emaktel/fs-api
servers:
  - url: http://localhost:37274
    description: Local development

# ---------------------------------------------------------------------------
# Security
# ---------------------------------------------------------------------------
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: >
        Required for remote requests when FSAPI_AUTH_TOKENS is configured.
        Localhost requests bypass authentication.

  # -------------------------------------------------------------------------
  # Headers
  # -------------------------------------------------------------------------
  parameters:
    XAllowedContexts:
      name: X-Allowed-Contexts
      in: header
      required: false
      description: >
        Comma-separated list of allowed FreeSWITCH contexts (domains).
        Use `*` for unrestricted access. Omit for backward-compatible
        unrestricted access.
      schema:
        type: string
        example: "customer1.example.com,customer2.example.com"
...
    CallUUID:
      name: uuid
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Call leg UUID
...
  headers:
    X-Request-ID:
      description: Unique identifier for the request, useful for log correlation
      schema:
        type: string
        format: uuid

  # -------------------------------------------------------------------------
  # Schemas
  # -------------------------------------------------------------------------
  schemas:
    SuccessMessage:
      type: object
      properties:
        status:
          type: string
          example: success
        message:
          type: string
      required: [status, message]
    ErrorMessage:
      type: object
      properties:
        status:
          type: string
          example: error
        message:
          type: string
      required: [status, message]
...
    ConferenceRequest:
      type: object
      required: [destination]
      properties:
        destination:
          type: string
          pattern: "^[0-9+*#]{1,20}$"
          description: Extension or external number to add to the call's conference room
        tollAllow:
          type: string
          pattern: "^[A-Za-z0-9_-]{1,32}(?:,[A-Za-z0-9_-]{1,32}){0,31}$"
          description: >
            The originating extension's toll classes, read server-side by the caller (the
            SvelteKit softphone reads them from the user's own extension; the public calls_v2
            proxy never forwards one), so the loopback dial is gated by the same dialplan toll
            rules a normal outbound call gets. Empty for internal-extension adds.
...
  responses:
    BadRequest:
      description: Invalid request
      headers:
        X-Request-ID:
          $ref: "#/components/headers/X-Request-ID"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorMessage"
...
    Forbidden:
      description: Domain / context not in allowed list
      headers:
        X-Request-ID:
          $ref: "#/components/headers/X-Request-ID"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorMessage"
    NotFound:
      description: Resource not found
      headers:
        X-Request-ID:
          $ref: "#/components/headers/X-Request-ID"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorMessage"
    BadGateway:
      description: FreeSWITCH ESL command error
      headers:
        X-Request-ID:
          $ref: "#/components/headers/X-Request-ID"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorMessage"
    ServiceUnavailable:
      description: FreeSWITCH ESL connection unavailable
      headers:
        X-Request-ID:
          $ref: "#/components/headers/X-Request-ID"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorMessage"

# ===========================================================================
# Paths
# ===========================================================================
paths:
...
  /v1/calls/{uuid}/conference:
    post:
      tags: [Calls]
      summary: Add a participant to the call as a server-mixed conference ("add people")
      description: >
        On the first add, moves the call and its bridged partner (read from the call's own
        channel dump; it must be the same tenant, 403 otherwise, 409 if it has hung up) into a
        per-call room `sp-<uuid>` on the `softphone` conference profile, one `uuid_transfer` per
        leg, then dials `destination` into it via a loopback through the call's own dialplan
        context; later adds reuse the room. A call in any other conference answers 409. The join
        is verified and a 503 is returned if the profile is unavailable. The dial is
        non-blocking — a 200 means the participant was queued, not yet connected.
      operationId: addToConference
      parameters:
        - $ref: "#/components/parameters/CallUUID"
        - $ref: "#/components/parameters/XAllowedContexts"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConferenceRequest"
      responses:
        "200":
          description: Participant queued into the conference
          headers:
            X-Request-ID:
              $ref: "#/components/headers/X-Request-ID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessMessage"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: The bridged partner has hung up, or the call is in a conference fs-api didn't start
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorMessage"
        "502":
          $ref: "#/components/responses/BadGateway"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
...